funcional

+++ src/stores/clubsStore.ts
@@ -1,5 +1,7 @@
 import { defineStore } from 'pinia';
 import axios from 'axios';
+import { useLoginStore } from '@/stores/milogin';
+
 const API_BASE_URL = 'http://localhost:3000';
 
 export const useClubsStore = defineStore('clubs', {
@@ -16,6 +18,11 @@ export const useClubsStore = defineStore('clubs', {
       }
     },
     async createClub(club: { name: string; description: string }) {
+      const loginStore = useLoginStore();
+      if (loginStore.user?.role !== 'admin') {
+        console.error('Acción no autorizada: El usuario no es admin');
+        return;
+      }
       try {
         const response = await axios.post(`${API_BASE_URL}/clubs`, club);
         this.clubs.push(response.data);
@@ -24,6 +31,11 @@ export const useClubsStore = defineStore('clubs', {
       }
     },
     async deleteClub(id: string) {
+      const loginStore = useLoginStore();
+      if (loginStore.user?.role !== 'admin') {
+        console.error('Acción no autorizada: El usuario no es admin');
+        return;
+      }
       try {
         await axios.delete(`${API_BASE_URL}/clubs/${id}`);
         this.clubs = this.clubs.filter(club => club._id !== id);
